+++ src/ng2-ui-auth.module.ts
@@ -13,33 +13,19 @@ import { JwtInterceptor } from './interceptor.service';
 import { StorageService, BrowserStorageService } from './storage.service';
 import { InjectionToken, ModuleWithProviders, NgModule, Type } from '@angular/core';
 import { HTTP_INTERCEPTORS, HttpClientModule, HttpClient } from '@angular/common/http';
+import { StorageType } from './storage-type.enum';
 
 @NgModule({ imports: [HttpClientModule] })
 export class Ng2UiAuthModule {
-    static forRootWithoutOptions(): ModuleWithProviders {
+    static forRoot(configOptions?: IPartialConfigOptions, defaultJwtInterceptor = true): ModuleWithProviders {
         return {
             ngModule: Ng2UiAuthModule,
             providers: [
+                ...configOptions ? [{ provide: CONFIG_OPTIONS, useValue: configOptions }] : [],
                 { provide: ConfigService, useClass: ConfigService, deps: [CONFIG_OPTIONS] },
                 { provide: StorageService, useClass: BrowserStorageService, deps: [ConfigService] },
                 { provide: SharedService, useClass: SharedService, deps: [StorageService, ConfigService] },
-                { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true, deps: [SharedService, ConfigService] },
-                { provide: OauthService, useClass: OauthService, deps: [HttpClient, SharedService, ConfigService, PopupService] },
-                { provide: PopupService, useClass: PopupService, deps: [ConfigService] },
-                { provide: LocalService, useClass: LocalService, deps: [HttpClient, SharedService, ConfigService] },
-                { provide: AuthService, useClass: AuthService, deps: [SharedService, LocalService, OauthService] },
-            ],
-        };
-    }
-    static forRoot(options: IPartialConfigOptions): ModuleWithProviders {
-        return {
-            ngModule: Ng2UiAuthModule,
-            providers: [
-                { provide: CONFIG_OPTIONS, useValue: options },
-                { provide: ConfigService, useClass: ConfigService, deps: [CONFIG_OPTIONS] },
-                { provide: StorageService, useClass: BrowserStorageService, deps: [ConfigService] },
-                { provide: SharedService, useClass: SharedService, deps: [StorageService, ConfigService] },
-                { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true, deps: [SharedService, ConfigService] },
+                ...defaultJwtInterceptor ? [{ provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true, deps: [SharedService, ConfigService] }] : [],
                 { provide: OauthService, useClass: OauthService, deps: [HttpClient, SharedService, ConfigService, PopupService] },
                 { provide: PopupService, useClass: PopupService, deps: [ConfigService] },
                 { provide: LocalService, useClass: LocalService, deps: [HttpClient, SharedService, ConfigService] },
@@ -62,4 +48,5 @@ export {
     JwtInterceptor,
     CONFIG_OPTIONS,
     IProviders,
+    StorageType,
 };