+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-#### 8.0.0-beta.6
-* added storage type 'none'
+#### 8.0.0-beta.7
+* added storage type 'none' and 'memmory' and added StorageType enum.
+* `forRoot` without options will give you the ability to inject your own config
+* you can decide to not inject the JwtInterceptor view `forRoot` function
 
 #### 8.0.0-beta.5
 * fix signup logged in issue